refactor[#153] : Slider에서 ImgSlider로 네이밍 변경
