+++ src/shared/ui/imgSlider/index.ts
@@ -1,0 +1,1 @@
+export { default as ImgSlider } from "./ImgSlider";
