--- src/shared/ui/slider/index.ts
@@ -1,1 +1,0 @@
-export { default as Slider } from "./Slider";
